adding documentation and arcitecture
