+++ app/src/main/java/com/example/clicker/network/domain/TwitchAuthentication.kt
@@ -1,6 +1,6 @@
 package com.example.clicker.network.domain
 
-import com.example.clicker.network.models.ValidatedUser
+import com.example.clicker.network.models.twitchAuthentication.ValidatedUser
 import com.example.clicker.util.Response
 import kotlinx.coroutines.flow.Flow
