--- app/src/main/java/com/example/clicker/network/models/FollowedLiveStreams.kt
@@ -1,5 +1,0 @@
-package com.example.clicker.network.models
-
-data class FollowedLiveStreams(
-    val data: List<StreamData>
-)
