+++ app/src/main/java/com/example/clicker/network/models/twitchRepo/FollowedLiveStreams.kt
@@ -1,0 +1,7 @@
+package com.example.clicker.network.models.twitchRepo
+
+import com.example.clicker.network.models.twitchRepo.StreamData
+
+data class FollowedLiveStreams(
+    val data: List<StreamData>
+)
