+++ app/src/test/java/com/example/clicker/network/TwitchRepoImplTest.kt
@@ -1,8 +1,8 @@
 package com.example.clicker.network
 
 import com.example.clicker.network.domain.TwitchRepo
-import com.example.clicker.network.models.FollowedLiveStreams
-import com.example.clicker.network.models.StreamData
+import com.example.clicker.network.models.twitchRepo.FollowedLiveStreams
+import com.example.clicker.network.models.twitchRepo.StreamData
 import com.example.clicker.network.repository.TwitchRepoImpl
 import com.example.clicker.util.Response
 import com.google.gson.Gson
